moved ndflattener.hpp to Common/include/toolboxes
